refactor: EnclaveEngine -> StateRuntimeEnclaveEngine
