+++ example/erc20/enclave/src/ecalls.rs
@@ -4,7 +4,7 @@ use anonify_ecall_types::cmd::*;
 use anonify_enclave::{context::AnonifyEnclaveContext, workflow::*};
 use anyhow::anyhow;
 use frame_common::crypto::NoAuth;
-use frame_enclave::{register_ecall, EnclaveEngine};
+use frame_enclave::{register_ecall, StateRuntimeEnclaveEngine};
 use std::{ptr, vec::Vec};
 
 register_ecall!(
