+++ example/key-vault/enclave/src/ecalls.rs
@@ -1,5 +1,5 @@
 use crate::ENCLAVE_CONTEXT;
-use frame_enclave::{register_ecall, EnclaveEngine};
+use frame_enclave::{register_ecall, StateRuntimeEnclaveEngine};
 use key_vault_ecall_types::cmd::*;
 use key_vault_enclave::{context::KeyVaultEnclaveContext, workflow::*};
 use std::{ptr, vec::Vec};
