+++ modules/anonify-enclave/src/commands/treekem.rs
@@ -7,7 +7,7 @@ use frame_common::{
     state_types::StateType,
     AccessPolicy,
 };
-use frame_enclave::EnclaveEngine;
+use frame_enclave::StateRuntimeEnclaveEngine;
 use frame_runtime::traits::*;
 use serde::{Deserialize, Serialize};
 use std::marker::PhantomData;
@@ -19,7 +19,7 @@ pub struct CommandByTreeKemSender<AP: AccessPolicy> {
     user_id: Option<AccountId>,
 }
 
-impl<AP> EnclaveEngine for CommandByTreeKemSender<AP>
+impl<AP> StateRuntimeEnclaveEngine for CommandByTreeKemSender<AP>
 where
     AP: AccessPolicy,
 {
@@ -96,7 +96,7 @@ pub struct CommandByTreeKemReceiver<AP> {
     ap: PhantomData<AP>,
 }
 
-impl<AP> EnclaveEngine for CommandByTreeKemReceiver<AP>
+impl<AP> StateRuntimeEnclaveEngine for CommandByTreeKemReceiver<AP>
 where
     AP: AccessPolicy,
 {
