+++ modules/anonify-enclave/src/context.rs
@@ -17,7 +17,7 @@ use frame_common::{
 #[cfg(feature = "backup-enable")]
 use frame_config::KEY_VAULT_ENCLAVE_MEASUREMENT;
 use frame_config::{ANONIFY_PARAMS_DIR, CMD_DEC_SECRET_DIR, IAS_ROOT_CERT};
-use frame_enclave::EnclaveEngine;
+use frame_enclave::StateRuntimeEnclaveEngine;
 #[cfg(feature = "backup-enable")]
 use frame_mra_tls::{
     key_vault::{
@@ -441,7 +441,7 @@ pub struct GetState<AP: AccessPolicy> {
     ecall_input: input::GetState<AP>,
 }
 
-impl<AP: AccessPolicy> EnclaveEngine for GetState<AP> {
+impl<AP: AccessPolicy> StateRuntimeEnclaveEngine for GetState<AP> {
     type EI = SodiumCiphertext;
     type EO = output::ReturnState;
 
@@ -485,7 +485,7 @@ pub struct GetUserCounter<AP: AccessPolicy> {
     ecall_input: input::GetUserCounter<AP>,
 }
 
-impl<AP: AccessPolicy> EnclaveEngine for GetUserCounter<AP> {
+impl<AP: AccessPolicy> StateRuntimeEnclaveEngine for GetUserCounter<AP> {
     type EI = SodiumCiphertext;
     type EO = output::ReturnUserCounter;
 
@@ -519,7 +519,7 @@ impl<AP: AccessPolicy> EnclaveEngine for GetUserCounter<AP> {
 #[derive(Debug, Clone, Default)]
 pub struct ReportRegistration;
 
-impl EnclaveEngine for ReportRegistration {
+impl StateRuntimeEnclaveEngine for ReportRegistration {
     type EI = input::Empty;
     type EO = output::ReturnRegisterReport;
